requirements.txt
fastapi
uvicorn
python-dotenv
pipecat-ai[webrtc,deepgram,google,openai,daily,silero]
pipecat-ai-small-webrtc-prebuilt
